Move utils back to routes folder I realize some other util functions I want to make will be shared across routers, so I brought it back to the parent folder haha
